arm: more fixes for ARM - RamFs supports STATIC_DRIVERS now, and reverted types back to long again now that the compiler is fixed.

+++ src/system/include/processor/arm_common/types.h
@@ -32,9 +32,9 @@ typedef signed short ARMint16_t;
 /** Define an 16bit unsigned integer type */
 typedef unsigned short ARMuint16_t;
 /** Define a 32bit signed integer type */
-typedef signed int ARMint32_t;
+typedef signed long ARMint32_t;
 /** Define a 32bit unsigned integer type */
-typedef unsigned int ARMuint32_t;
+typedef unsigned long ARMuint32_t;
 /** Define a 64bit signed integer type */
 typedef signed long long ARMint64_t;
 /** Define a 64bit unsigned integer type */
